backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Pydantic
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.26.0

# LangGraph & AI
langgraph>=0.2.0,<0.3.0
langchain>=0.2.0,<0.3.0
langchain-anthropic>=0.1.0,<0.2.0
langchain-core>=0.2.0,<0.3.0

# Email
sendgrid==6.11.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=22.0.0
